move IdentityGetter and ReferenceGetter to value_getter.hpp; add annotation
